add(hooks): add useParseAddress hook
useParseAddress hook takes all the logic for the address parsing in the
address form and extracts into a hook.
