+++ src/app/user/create-labels/components/AddressForm.tsx
@@ -1,52 +1,9 @@
 import { formInputs, statesList } from "~/lib/lists";
 import { type FormData, type HandleChange } from "~/lib/definitions";
-import { useState } from "react";
-import getParsedAddress from "~/lib/parseAddress";
-import { AddressParsingError } from "~/lib/customErrors";
-import * as Sentry from "@sentry/nextjs";
+import useParseAddress from "~/hooks/useParseAddress";
 
-export default function AddressForm({ formData, handleChange, label }: { formData: FormData; handleChange: HandleChange; label: string }) {
-  const [pastedAddress, setPastedAddress] = useState("");
-  const [parsingErrorMessage, setParsingErrorMessage] = useState("");
-
-  const parseAddress = async (address: string) => {
-    try {
-      const formattedAddress = await getParsedAddress(address);
-
-      if (label === "From") {
-        handleChange({
-          FromStreet: formattedAddress.address,
-          FromAddress: formattedAddress.addressTwo === " " ? "" : formattedAddress.addressTwo,
-          FromCity: formattedAddress.city,
-          FromState: formattedAddress.state,
-          FromZip: formattedAddress.zipcode,
-        });
-      } else if (label === "To") {
-        handleChange({
-          ToStreet: formattedAddress.address,
-          ToAddress: formattedAddress.addressTwo === " " ? "" : formattedAddress.addressTwo,
-          ToCity: formattedAddress.city,
-          ToState: formattedAddress.state,
-          ToZip: formattedAddress.zipcode,
-        });
-      }
-
-      setPastedAddress("");
-      setParsingErrorMessage("");
-    } catch (err) {
-      /**
-       * * If error originated from bad request to smarty API, forward to sentry
-       * * Otherwise just show the user a parsing error
-       */
-      if (err instanceof AddressParsingError) {
-        setParsingErrorMessage(err.message);
-      } else {
-        Sentry.captureException(err);
-        setParsingErrorMessage("Could not parse address");
-      }
-    }
-  };
-
+export default function AddressForm({ formData, handleChange, label }: { formData: FormData; handleChange: HandleChange; label: "From" | "To" }) {
+  const { pastedAddress, setPastedAddress, parsingErrorMessage, parseAddress } = useParseAddress(handleChange, label);
   return (
     <section className="rounded-2xl bg-linear-gradient">
       <div className="flex h-[calc(100%-3px)] w-[calc(100%-3px)] translate-x-[1.5px] translate-y-[1.5px] flex-col justify-between gap-8 rounded-2xl bg-radial-gradient p-5">
